retained: only the UI part in binary and moved terminal to output lib
